rename app to api in example_project
