+++ example_project/api/views.py
@@ -7,8 +7,8 @@ from rest_framework.views import APIView
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
-from app import serializers
-from app.models import Person
+from api import serializers
+from api.models import Person
 
 logger = getLogger('django')
 api_logger = APILoggingDecorator(logger=logger, level='INFO')
