--- example_project/app/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class AppConfig(AppConfig):
-    name = 'app'
